+++ src/components/TypeaheadSelect/useTypeaheadSelect.ts
@@ -10,7 +10,10 @@ export type UseTypeaheadSelectOptions<
   Values extends FieldValues,
 > = Pick<
   TypeaheadSelectProps<DataItem, Values>,
-  'debounceTime' | 'onDebouncedChange' | 'options'
+  | 'debounceTime'
+  | 'defaultShowDropdown'
+  | 'isQueryLoading'
+  | 'onDebouncedChange'
 >;
 
 export const useTypeaheadSelect = <
@@ -19,17 +22,20 @@ export const useTypeaheadSelect = <
   Values extends FieldValues,
 >({
   debounceTime,
+  defaultShowDropdown,
+  isQueryLoading,
   onDebouncedChange,
-  options,
 }: UseTypeaheadSelectOptions<DataItem, Values>) => {
   const [selectedItems, setSelectedItems] = useState<DataItem[]>([]);
-  const [showDropdown, setShowDropdown] = useState(false);
+  const [showDropdown, setShowDropdown] = useState(
+    defaultShowDropdown ?? false,
+  );
   const dropdownRef = useRef<DropdownElement>(null);
   const searchInputRef = useRef<HTMLInputElement | null>(null);
-  const { isLoading, query, setQuery } = useTypeaheadQuery<DataItem>({
+  const { isLoading, query, setQuery } = useTypeaheadQuery({
     debounceTime,
+    isQueryLoading,
     onDebouncedChange,
-    options,
   });
   const clearSelectedItem = (index: number) =>
     setSelectedItems(items =>
@@ -37,7 +43,7 @@ export const useTypeaheadSelect = <
     );
   useOutsideClick(dropdownRef, () => setShowDropdown(false));
   useEffect(() => {
-    if (showDropdown) searchInputRef.current?.focus();
+    if (showDropdown) setTimeout(() => searchInputRef.current?.focus(), 50);
     else setQuery('');
   }, [showDropdown, setQuery]);
   return {